% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/variable.R
\name{as.rowvar}
\alias{as.rowvar}
\title{A class for variables}
\usage{
as.rowvar(units)
}
\arguments{
\item{units}{}
}
\description{
Specify a named variable with units
}
\examples{
distance <- as.rowvar("cm")
}

